Add OpenCommand to open GitHub URLs
Tutors have no way of accessing the student's github directly.

This wastes a lot of time for the tutors to manually input
the student's respective github link in their browsers.

Let's:
- Implement OpenCommand to enable tutors to open GitHub profiles
efficiently via the CLI.
- Implement a BrowserService interface to allow for stubs
for easier testing without opening tabs in the environment
- Add tests to cover for the new command.

This approach decouples the URL opening logic and allows for flexibility
and easy testing, while enhancing the user experience by providing a
simple command to access GitHub profiles directly.