Implement assignment related functionality
Assignments are not tracked, only indicator is progress attribute.

Assignments with deadlines needs to be easily tracked by TAs.

Lets:
- Add a TimedEvent abstract class.
- Add assignment class to be able to create Assignments which
inherits from the TimedEvent superclass
- Add assignment related commands to assign assignments to
classes and individuals with deadlines
- Add test coverage to cover for added functionalities

This ensures that assignments can be tracked easily.
This also is open for extension when other deadline types
may be added in the future.